% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testnslv.R
\name{print.test.nleqslv}
\alias{print.test.nleqslv}
\alias{print}
\title{Printing the result of \code{testnslv}}
\usage{
\method{print}{test.nleqslv}(x, digits = 4, width.cutoff = 45L, ...)
}
\arguments{
\item{x}{a \code{test.nleqslv} object}

\item{digits}{specifies the minimum number of significant digits to be
printed in values.}

\item{width.cutoff}{integer passed to \code{\link{deparse}} which sets the
cutoff at which line-breaking is tried.}

\item{\dots}{additional arguments to \code{print}.}
}
\value{
It returns the object \code{x} invisibly.
}
\description{
Print method for \code{test.nleqslv} objects.
}
\details{
This is the \code{print} method for objects inheriting from class
\code{test.nleqslv}. It prints the call to \code{testnslv} followed by the
description of the experiment (if the \code{title} argument was specified in
the call to \code{testnslv}) and the dataframe containing the results of
\code{testnslv}.
}
\examples{

dslnex <- function(x) {
    y <- numeric(2)
    y[1] <- x[1]^2 + x[2]^2 - 2
    y[2] <- exp(x[1]-1) + x[2]^3 - 2
    y
}
xstart <- c(1.5,0.5)
fstart <- dslnex(xstart)
z <- testnslv(xstart,dslnex)
print(z)

}
\keyword{print}
